alternative solution strategy: "forcing"
this is what we usually do on paper, we're searching for fields that
must have certain values and then look whether these new fields force
other fields.

it's implemented by generating all possibilities for a row/column
(considering the already known fields) and looking for "common" fields
in those possibilities. if a field has the same values in all possible
bindings, then we know it must have the value we found.

this isn't finished yet, as we have to rerun it by hand for now.